Modification : update tests with refactoring in several modules
