+++ oc_lettings_site/tests/tests_url.py
@@ -1,0 +1,12 @@
+"""
+Tests module for oc_lettings_site app urls
+"""
+from django.urls import reverse, resolve
+
+
+class TestOcLettingsSiteUrl:
+    def test_index_url(self):
+        path = reverse("index")
+
+        assert path == "/"
+        assert resolve(path).view_name == 'index'
